Introduce network tab to establish connection
Moves the code that's not about signalling and RTC connection setup
out of webrtc_signaling.rs and into network.rs. This is also where the
new view lives.

This split also enabled multiple WebRTC connections! Either with the
same peer (not the intention) or completely different  peers.

To make enough UI space, tabs are added, for now controlled at
the bottom. All the UI is still horrible and will need to be cleaned up.